[StationHistoricoView] Configuracion de navegacion para acceder a la vista de datos historicos

+++ app/src/main/java/ar/edu/ort/respirar/data/dto/station/Pm1.kt
@@ -2,5 +2,5 @@ package ar.edu.ort.respirar.data.dto.station
 
 data class Pm1(
     var type: String?,
-    var value: Int?
+    var value: Double?
 )
